requirements-examples.txt
# Pin for reproducible examples; relax as needed.
door-stabilizer>=0.4.1
